requirements.txt
# Data Processing & Analysis
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Big Data Processing
pyspark==3.5.1
kafka-python==2.0.2

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Visualization & Dashboard
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2 